// lib/generated/inputTypeSchemas/ActivityOrderByWithRelationInputSchema.ts
import type { Prisma } from '../../prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';
import { SortOrderInputSchema } from './SortOrderInputSchema';
import { HotelOrderByWithRelationInputSchema } from './HotelOrderByWithRelationInputSchema';
import { ActivityOrderByRelevanceInputSchema } from './ActivityOrderByRelevanceInputSchema';

export const ActivityOrderByWithRelationInputSchema: z.ZodType<Prisma.ActivityOrderByWithRelationInput> = z.object({
  id: z.lazy(() => SortOrderSchema).optional(),
  name: z.lazy(() => SortOrderSchema).optional(),
  name_esp: z.lazy(() => SortOrderSchema).optional(),
  description: z.lazy(() => SortOrderSchema).optional(),
  description_esp: z.lazy(() => SortOrderSchema).optional(),
  image: z.lazy(() => SortOrderSchema).optional(),
  profileImage: z.union([ z.lazy(() => SortOrderSchema),z.lazy(() => SortOrderInputSchema) ]).optional(),
  hotelId: z.lazy(() => SortOrderSchema).optional(),
  type: z.lazy(() => SortOrderSchema).optional(),
  start: z.lazy(() => SortOrderSchema).optional(),
  end: z.lazy(() => SortOrderSchema).optional(),
  price: z.lazy(() => SortOrderSchema).optional(),
  cashOnly: z.lazy(() => SortOrderSchema).optional(),
  freeOfCharge: z.lazy(() => SortOrderSchema).optional(),
  location: z.lazy(() => SortOrderSchema).optional(),
  isRecurring: z.lazy(() => SortOrderSchema).optional(),
  recurrencePattern: z.lazy(() => SortOrderSchema).optional(),
  recurrenceDays: z.lazy(() => SortOrderSchema).optional(),
  recurrenceEndDate: z.lazy(() => SortOrderSchema).optional(),
  createdAt: z.lazy(() => SortOrderSchema).optional(),
  updatedAt: z.lazy(() => SortOrderSchema).optional(),
  Hotel: z.lazy(() => HotelOrderByWithRelationInputSchema).optional(),
  _relevance: z.lazy(() => ActivityOrderByRelevanceInputSchema).optional()
}).strict();

export default ActivityOrderByWithRelationInputSchema;